% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codon_to_AA.R
\name{codon_to_AA}
\alias{codon_to_AA}
\title{codon_to_AA}
\usage{
codon_to_AA(codons)
}
\arguments{
\item{codons}{vector of codon indices to be translated to amino acids}
}
\value{
function returns a amino acid sequence
}
\description{
codon_to_AA
}
\examples{
codon_to_AA(c("UUU", "UUC", "UUA"))
}
