% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_function.R
\name{dna_1}
\alias{dna_1}
\title{Function three}
\usage{
dna_1(dna_2, start = 1)
}
\arguments{
\item{start}{description of parameter}

\item{dna}{description of parameter}
}
\value{
description of function output
}
\description{
Function three
}
