% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_function.R
\name{extract_codons}
\alias{extract_codons}
\title{Function three}
\usage{
extract_codons(RNA_sequence, start = 1)
}
\arguments{
\item{RNA_sequence}{gets extracted into codons of three nucleotide each}
}
\value{
strings of different codons each with three nucleotides from RNA sequence
}
\description{
Function three
}
\examples{
RNA_sequence <- "ACGUUGCA"
extract_codons(RNA_sequence)
}
