% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleotide_sequence_sampler.R
\name{nucleotide_sequence_sampler}
\alias{nucleotide_sequence_sampler}
\title{nucleotide_sequence_sampler}
\usage{
nucleotide_sequence_sampler(sequence_length)
}
\arguments{
\item{sequence_length}{contains length of sample sequence.}
}
\value{
function returns a sampled nucleotide string.
}
\description{
nucleotide_sequence_sampler
}
\examples{
example_sequence <- nucleotide_sequence_sampler(100)
print(example_sequence)
}
