% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T_to_U_converter.R
\name{T_to_U_converter}
\alias{T_to_U_converter}
\title{T_to_U_converter}
\usage{
T_to_U_converter(dna_sequence)
}
\arguments{
\item{dna_sequence}{sequence of nucleotides containing letters "A", "C", "G" and "T" - a DNA sequence}
}
\value{
sequence of nucleotides containing letters "A", "C", "G" and "U" - a RNA sequence
}
\description{
T_to_U_converter
}
\examples{
T_to_U_converter("ACGTTGCA")
}
