% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amino_acid_plot.R
\name{amino_acid_plot}
\alias{amino_acid_plot}
\title{Amino acid frequency plot}
\usage{
amino_acid_plot(aa_sequence)
}
\arguments{
\item{aa_sequence}{An amino acid sequence in a string}
}
\value{
A plot of the frequency of the different amino acids present in the sequence
}
\description{
Amino acid frequency plot
}
\examples{
INS <- "MALWMRLLPLLALLALWGPDPAAAFVNQHLCGSHLVEALYLVCGERGFFYTPKTRREAEDLQVGQVELGGGPGAGSLQPLALEGSLQKRGIVEQCCTSICSLYQLENYCN"
amino_acid_plot(INS)
entH <- "MINKIKILFSFLALLLSFTSYAKAEDLHDKSELTDLALANAYGQYNHPFIKENIKSDEISGEKDLIFRNQGDSGNDLRVKFATADLAQKFKNKNVDIYGASFYYKCEKISENISECLYGGTTLNSEKLAQERVIGANVWVDGIQKETELIRTNKKNVTLQELDIKIRKILSDKYKIYYKDSEISKGLIEFDMKTPRDYSFDIYDLKGENDYEIDKIYEDNKTLKSDDISHIDVNLYTKKKV"
amino_acid_plot(entH)
}
