% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codons_function.R
\name{extract_codons}
\alias{extract_codons}
\title{Function three}
\usage{
extract_codons(sequence, start = 1)
}
\arguments{
\item{start}{description of parameter}

\item{dna}{description of parameter}
}
\value{
description of function output
}
\description{
Function three
}
